+++ js/sort-pictures.js
@@ -1,9 +1,6 @@
-import { closePictureEditor, setOnFormSubmit, openPictureLoadEditor } from './load-picture-editor.js';
-import { loadBigPicture } from './open-full-picture.js';
+import { setOnFormSubmit } from './load-picture-editor.js';
 import { renderPictures } from './pictures.js';
-import { getData, sendData } from './api.js';
-import { debounce, showAlert } from './utils.js';
-import { showSuccessMessage, showErrorMessage } from './message.js';
+import { debounce } from './utils.js';
 
 const RANDOM_PICTURES_LENGTH = 10;
 const RERENDERER_DELAY = 500;
@@ -15,7 +12,11 @@ const filterDiscussedButton = document.querySelector('#filter-discussed');
 const similarListElemets = document.querySelector('.pictures');
 const sortButtons = document.querySelectorAll('.img-filters__button');
 const childSimilarElements = similarListElemets.children;
-let data;
+
+const showSortButtons = () => {
+  filtersBlock.classList.remove('img-filters--inactive');
+};
+
 
 const changeSortClass = (evt) => {
   const activeClass = 'img-filters__button--active';
@@ -38,13 +39,13 @@ const deletePreviousPictures = () => {
 };
 
 // по умолч
-const sortDefault = (debounce(() => {
+const sortDefault = (debounce((data) => {
   deletePreviousPictures();
   renderPictures(data);
 }, RERENDERER_DELAY));
 
 // случайные
-const sortRandomData = (debounce(() => {
+const sortRandomData = (debounce((data) => {
   deletePreviousPictures();
   const randomData = data
     .slice()
@@ -54,7 +55,7 @@ const sortRandomData = (debounce(() => {
 }, RERENDERER_DELAY));
 
 // по убыванию кол-ва комментов
-const sortComments = (debounce(() => {
+const sortComments = (debounce((data) => {
   deletePreviousPictures();
   const commentsData = data
     .slice()
@@ -62,38 +63,13 @@ const sortComments = (debounce(() => {
   renderPictures(commentsData);
 }, RERENDERER_DELAY));
 
-const getDefaultData = async () => {
-  try {
-    data = await getData();
-    renderPictures(data);
-    loadBigPicture(data);
-    filtersBlock.classList.remove('img-filters--inactive');
-    openPictureLoadEditor();
-  } catch (err) {
-    showAlert(err.message);
-  }
-};
-
-setOnFormSubmit(async (pictures) => {
-  try {
-    await sendData(pictures);
-    closePictureEditor();
-    showSuccessMessage();
-    loadBigPicture(data);
-  } catch {
-    showErrorMessage();
-  }
-});
-
-const sortPictures = () => {
+const sortPictures = (data) => {
   sortButtons.forEach((button) => {
     button.addEventListener('click', changeSortClass);
   });
-  filterDiscussedButton.addEventListener('click', sortComments);
-  filterRandomButton.addEventListener('click', sortRandomData);
-  filterDefaultButton.addEventListener('click', sortDefault);
+  filterDiscussedButton.addEventListener('click', () => sortComments(data));
+  filterRandomButton.addEventListener('click', () => sortRandomData(data));
+  filterDefaultButton.addEventListener('click', () => sortDefault(data));
 };
 
-export { sortPictures, getDefaultData, setOnFormSubmit };
-
-// доделать: 1. комменты криво при сортировке.
+export { sortPictures, setOnFormSubmit, showSortButtons };
